# Made cachedEndpoints read-only to avoid write+read races