Fixes https://github.com/zalando/skipper/issues/2079

Signed-off-by: Roman Zavodskikh <roman.zavodskikh@zalando.de>