@RomanZavodskikh 2022-09-26
Need to write some performance tests to find out whether we need to remove cachedEndpoints.