@RomanZavodskikh 2022-10-06
New benchmark results:
```
benchmark                        old ns/op     new ns/op     delta
BenchmarkCachedEndpoints1-16     47.8          1901          +3876.99%
BenchmarkCachedEndpoints2-16     49.2          4664          +9371.97%
BenchmarkCachedEndpoints3-16     49.6          4866          +9702.58%

benchmark                        old allocs     new allocs     delta
BenchmarkCachedEndpoints1-16     0              17             +Inf%
BenchmarkCachedEndpoints2-16     0              17             +Inf%
BenchmarkCachedEndpoints3-16     0              17             +Inf%

benchmark                        old bytes     new bytes     delta
BenchmarkCachedEndpoints1-16     0             784           +Inf%
BenchmarkCachedEndpoints2-16     0             784           +Inf%
BenchmarkCachedEndpoints3-16     0             784           +Inf%
```