@AlexanderYastrebov 2022-10-11
> right now I lean towards merging only first commit 

Performance tests added by https://github.com/zalando/skipper/pull/2092